src/client/requirements.txt
websocket-client
requests
pyautogui
pillow
keyboard
psutil
opencv-python
pytest
